+++ packages/typescript/lib/quickstart/createLanguageServicePlugin.ts
@@ -1,7 +1,7 @@
 import { FileMap, Language, LanguagePlugin, createLanguage } from '@volar/language-core';
 import type * as ts from 'typescript';
 import { resolveFileLanguageId } from '../common';
-import { createProxyLanguageService } from '../node/decorateLanguageService';
+import { createProxyLanguageService } from '../node/proxyLanguageService';
 import { decorateLanguageServiceHost, searchExternalFiles } from '../node/decorateLanguageServiceHost';
 
 export const externalFiles = new WeakMap<ts.server.Project, string[]>();
@@ -60,12 +60,11 @@ export function createLanguageServicePlugin(
 						}
 					);
 
-					const languageServiceProxy = createProxyLanguageService(info.languageService);
-					languageServiceProxy.initialize(language);
+					const { proxy, initialize } = createProxyLanguageService(info.languageService);
+					info.languageService = proxy;
+					initialize(language);
 					decorateLanguageServiceHost(ts, language, info.languageServiceHost);
 					setup?.(language);
-
-					return languageServiceProxy.proxy;
 				}
 
 				return info.languageService;
